requirements.txt
google-api-python-client==1.12.8
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
lxml==4.6.2
openpyxl==3.0.6
requests==2.25.1
tqdm==4.57.0
xlrd==2.0.1
